+++ src/certificate.h
@@ -1,0 +1,36 @@
+/*
+ * IoTaaP Cloud Certificate
+ * 
+ * IoTaaP Team <contact@iotaap.io>
+ * https://www.iotaap.io
+ * 
+ */
+
+#ifndef __IOTAAP_CLOUD_CERTIFICATE_H__
+#define __IOTAAP_CLOUD_CERTIFICATE_H__
+
+const char *iotaap_mqtts_certificate =
+    "-----BEGIN CERTIFICATE-----\n"
+    "MIIDlTCCAn2gAwIBAgIUd6oLZtAfYIjLOg2908R3hxQZ3lUwDQYJKoZIhvcNAQEL\n"
+    "BQAwWjELMAkGA1UEBhMCSFIxEDAOBgNVBAgMB0Nyb2F0aWExDzANBgNVBAcMBlph\n"
+    "Z3JlYjEPMA0GA1UECgwGSW9UYWFQMRcwFQYDVQQDDA5tcXR0LmlvdGFhcC5pbzAe\n"
+    "Fw0yMDAzMDgxMTA3MzhaFw0yNTAzMDcxMTA3MzhaMFoxCzAJBgNVBAYTAkhSMRAw\n"
+    "DgYDVQQIDAdDcm9hdGlhMQ8wDQYDVQQHDAZaYWdyZWIxDzANBgNVBAoMBklvVGFh\n"
+    "UDEXMBUGA1UEAwwObXF0dC5pb3RhYXAuaW8wggEiMA0GCSqGSIb3DQEBAQUAA4IB\n"
+    "DwAwggEKAoIBAQCtKghvPBU2xnOXsx0W/5y1ew8AYuws5cSkzukgzTkQI0XTO0n5\n"
+    "IT8CQD86Qi5/Q5qQtw0BQMUi5DEAPv/pHI0KuAdOaWYqlzA7AguP/4ubHvyJCNve\n"
+    "Id6584YGLLw3YRSHAAFuMSwUd09NcLi21CwwSnH8K6GYhSDzYC/N9E6vzlpmSb8f\n"
+    "48FeFkawHMrBgA/OFN7ffmIXe3QoCl4Ui3czk/rG3GN4J0rwpt16H8nvbGymIjBp\n"
+    "9cmya8US9lFDbpuCDaKTlH7Dla6ibMkDpOA8JnwukJEwiaEJR5wN9n3NwY/JEVXa\n"
+    "SXyEtQTKAqPoZj1y3CSmq/Y8FfPz1IkwPzMNAgMBAAGjUzBRMB0GA1UdDgQWBBSZ\n"
+    "m4E30lB6RHa1slkKEmFUMbZFWDAfBgNVHSMEGDAWgBSZm4E30lB6RHa1slkKEmFU\n"
+    "MbZFWDAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4IBAQB7fWkOQ1Uj\n"
+    "4/W4LLkFCHa+OYZIy+t3r8bgYjKoCqH8+DlXZWpwThjn8a+heTeYCknbh/D8270k\n"
+    "edzBhUjK06qoDWpXOI4eL3+7vffLS+AEanAi/XuiqYi7spoMSz2JwXFjc/94HpYf\n"
+    "WC4fcelqIdQl0qR4dnRH9otKfz4KeqhgVcjL0nyPpt2Q9UndN0nT043bthW5kk35\n"
+    "f2eoPel0okXFNmcvusS1pEvg8ddfM6GkJTYZr5aj5ZjitRE0ZwGBayE76JkLrSlZ\n"
+    "fbPnrsnBM5OA7ytgR7LyFB4bHaJCM81NAYDQGG9hcsiaZeCEOQHIR97xd32a52Tm\n"
+    "QySj9+LAkqon\n"
+    "-----END CERTIFICATE-----\n";
+
+#endif