+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 # Change Log
+
+## [0.1.0] - 2020-03-08
+ 
+### Added
+
+First stable version. Status publishing, publish and device publish working.
+Added new IoTaaP Cloud certificate.
+Added some future functions templates.
+ 
+### Changed
+   
+### Fixed
      
 ## [0.0.1] - 2019-09-15